regrouped some properties in hierarchy of models
